requirements.txt
Pillow==9.4.0
requests==2.25.1